Refactoring for intermittent test failures
